nca-llm-main/requirements.txt
agentpy
company
imageio
matplotlib
networkx
numpy
openai
Pillow
producer
PyYAML
rewards
tiktoken
torch
openllm
ollama
openllm[vllm]
